Decompose InstrumentService into focused services
- Extract TransactionProfitService for profit/loss calculations (115 lines)
- Extract InstrumentSnapshotService for metrics aggregation (148 lines)
- Reduce InstrumentService from 307 to 62 lines (80% reduction)
- Add comprehensive tests for new services
- Update InstrumentServiceTest to verify delegation

Closes #982
